# REST APIs Implementation/api/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://todo.swagger.io/v2
  - url: http://todo.swagger.io/v2
...
paths:
  /api/films:
    post:
      tags:
        - films
      summary: Create a new film
      description: A new film is created by the authenticated user.
      operationId: createFilm
      requestBody:
        description: Representation of the film to be created
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Film"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: Films
  /api/films/private:
    get:
      tags:
        - films/private
      summary: Retrieve the private films of the logged-in user
      description:
        The private films of the logged-in user are retrieved. A pagination
        mechanism is implemented to limit the size of messages.
      operationId: getPrivateFilms
      parameters:
        - name: pageNo
          in: query
          description: "The id of the requested page (if absent, the first page is returned)"
          required: false
          style: form
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200"
        "401":
          description: Unauthorized
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: Filmsprivate
  /api/films/private/{filmId}:
    get:
      tags:
        - "films/private/{filmId}"
      summary: Retrieve a private film
      description:
        The private film with ID filmId is retrieved. This operation can
        be performed on the film if the user who performs the operation is the film's
        owner.
      operationId: getSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to retrieve
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmsprivatefilmId
    put:
      tags:
        - "films/private/{filmId}"
      summary: Update a private film
      description:
        The private film with ID filmId is updated. This operation does
        not allow changing its visibility. This operation can be performed only by
        the owner.
      operationId: updateSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to update
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Film"
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description: "Conflict (attempt to make unallowed change, e.g., visibility)"
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmsprivatefilmId
    delete:
      tags:
        - "films/private/{filmId}"
      summary: Delete a private film
      description:
        The private film with ID filmId is deleted. This operation can
        only be performed by the owner.
      operationId: deleteSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to delete
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmsprivatefilmId
  /api/films/public:
    get:
      tags:
        - films/public
      summary: Retrieve the public films
      description:
        "The public films (i.e., the films that are visible for all the\
        \ users of the service) are retrieved. This operation does not require authentication.\
        \ A pagination mechanism is implemented to limit the size of messages."
      operationId: getPublicFilms
      parameters:
        - name: pageNo
          in: query
          description: "The id of the requested page (if absent, the first page is returned)"
          required: false
          style: form
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200"
        "500":
          description: Internal Server Error
          content: {}
      x-swagger-router-controller: Filmspublic
  /api/films/public/invited:
    get:
      tags:
        - films/public/invited
      summary:
        "Retrieve the public films that the logged-in user has been invited\
        \ to review, possibly filtered by the review's invitation Status"
      description:
        "The public films that the logged-in user has been invited to review\
        \ are retrieved, possibly filtered by the review's invitation Status. A pagination\
        \ mechanism is implemented to limit the size of messages."
      operationId: getInvitedFilms
      parameters:
        - name: invitationStatus
          in: query
          description:
            "The invitation Status used to filter the reviews (if absent,\
            \ all reviews are considered)"
          required: false
          style: form
          explode: true
          schema:
            type: string
            enum:
              - PENDING
              - ACCEPTED
              - REFUSED
        - name: pageNo
          in: query
          description: "The id of the requested page (if absent, the first page is returned)"
          required: false
          style: form
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200"
        "401":
          description: Unauthorized
          content: {}
        "404":
          description: Not found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: Filmspublicinvited
  /api/films/public/{filmId}:
    get:
      tags:
        - "films/public/{filmId}"
      summary: Retrieve a public film
      description:
        The public film with ID filmId is retrieved. This operation does
        not require authentication.
      operationId: getSinglePublicFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to retrieve
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      x-swagger-router-controller: FilmspublicfilmId
    put:
      tags:
        - "films/public/{filmId}"
      summary: Update a public film
      description:
        The public film with ID filmId is updated. This operation does
        not allow changing its visibility. This operation can be performed only by
        the owner.
      operationId: updateSinglePublicFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to update
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Film"
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description: "Conflict (attempt to make unallowed change, e.g., visibility)"
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmId
    delete:
      tags:
        - "films/public/{filmId}"
      summary: Delete a public film
      description:
        The public film with ID filmId is deleted. This operation can only
        be performed by the owner.
      operationId: deleteSinglePublicFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to delete
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmId
  /api/films/public/{filmId}/reviews:
    get:
      tags:
        - "films/public/{filmId}/reviews"
      summary:
        "Retrieve the list of all the reviews that have been issued/completed\
        \ for a film, possibly filtered by the invitation Status if the user is the\
        \ film owner"
      description:
        "All the reviews that have been issued/completed for the film with\
        \ ID filmId are retrieved. If the requesting user is the owner of the film,\
        \ he also sees the reviews invitationStatus, otherwise it is not shown. A\
        \ pagination mechanism is implemented to limit the size of messages. This\
        \ operation does not require authentication."
      operationId: getFilmReviews
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: invitationStatus
          in: query
          description:
            "The invitation Status used to filter the reviews (if absent,\
            \ all reviews are considered). This query parameter only has effect for\
            \ the film owner."
          required: false
          style: form
          explode: true
          schema:
            type: string
            enum:
              - PENDING
              - ACCEPTED
              - REFUSED
        - name: pageNo
          in: query
          description: "ID of the requested page (if absent, the first page is returned)'"
          required: false
          style: form
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_1"
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      x-swagger-router-controller: FilmspublicfilmIdreviews
    post:
      tags:
        - "films/public/{filmId}/reviews"
      summary: Issue film review to some users
      description:
        The film with ID filmId is assigned to one or more users for review
        and the corresponding reviews are created. The users are specified in the
        review representations in the request body. This operation can only be performed
        by the owner.
      operationId: issueFilmReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description:
          "only the necessary parts of the new film reviews, including\
          \ the users to whom they are issued"
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Review_post"
        required: true
      responses:
        "207":
          description: Multi-Status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/inline_response_207"
                x-content-type: application/json
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description: "Conflict (e.g., wrong film id or unknown reviewer id)"
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviews
  /api/films/public/{filmId}/reviews/invitationStatus:
    get:
      tags:
        - "films/public/{filmId}/reviews/invitationStatus"
      summary:
        "Retrieve the invitationStatus for all reviews that has been issued/completed\
        \ for a film, possibly filtered"
      description:
        The invitationStatus for all the reviews of the film with ID filmID
        are retrieved. This operation can only be executed by the authenticated film
        owner.
      operationId: getReviewsInvitationStatus
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: invitationStatus
          in: query
          description:
            "The invitation Status used to filter the reviews (if absent,\
            \ all reviews are considered). This query parameter only has effect for\
            \ the film owner."
          required: false
          style: form
          explode: true
          schema:
            type: string
            enum:
              - PENDING
              - ACCEPTED
              - REFUSED
        - name: pageNo
          in: query
          description: "ID of the requested page (if absent, the first page is returned)'"
          required: false
          style: form
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_2"
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviewsinvitationStatus
  /api/films/public/{filmId}/reviews/{reviewerId}:
    get:
      tags:
        - "films/public/{filmId}/reviews/{reviewerId}"
      summary: Retrieve a review that has been issued/completed for a film
      description:
        "The review of the film with ID filmID issued to the user with\
        \ ID reviewerId is retrieved. This operation does not require authentication.\
        \ If it is called by the authenticated owner or reviewer, it also returns\
        \ the invitationStatus of the review."
      operationId: getSingleReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: reviewerId
          in: path
          description: ID of the user to whom the review has been issued
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_3"
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      x-swagger-router-controller: FilmspublicfilmIdreviewsreviewerId
    put:
      tags:
        - "films/public/{filmId}/reviews/{reviewerId}"
      summary: Complete a review
      description:
        "The review of the film with ID filmId and issued to the user with\
        \ ID reviewerId is completed. This operation only allows setting the \"completed\"\
        \ property to the \"true\" value, and changing the values of the \"reviewDate\"\
        , \"rating\", and \"review\" properties. This operation can be performed only\
        \ by the invited reviewer."
      operationId: updateSingleReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose review must be completed
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: reviewerId
          in: path
          description: ID of the user to whom the review has been issued
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: The properties of the Review object that must be updated
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review_public"
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviewsreviewerId
    delete:
      tags:
        - "films/public/{filmId}/reviews/{reviewerId}"
      summary: Delete a review invitation
      description:
        "The review of the film with ID filmId and issued to the user with\
        \ ID reviewerId is deleted. This operation can only be performed by the owner,\
        \ and only if the review has not yet been completed by the reviewer."
      operationId: deleteSingleReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose review invitation must be deleted
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: reviewerId
          in: path
          description: ID of the user to whom the review has been issued
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviewsreviewerId
    patch:
      tags:
        - "films/public/{filmId}/reviews/{reviewerId}"
      summary: Change invitation Status of a review
      description:
        Change the invitation Status of the review of the film with ID
        filmId and issued to the user with ID reviewerId. This operation allows setting
        the "invitationStatus" property of a "PENDING" review to "ACCEPTED" or "REFUSED".
        This operation can be performed only by the invited reviewer.
      operationId: updateSingleReviewInvitationStatus
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose review must be completed
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: reviewerId
          in: path
          description: ID of the user to whom the review has been issued
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: The properties of the Review object that must be updated
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review_invitationStatus"
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description:
            "Conflict (attempt to make unallowed change from ACCEPTED/REFUSED\
            \ review, or to PENDING)"
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviewsreviewerId
  /api/films/public/{filmId}/reviews/{reviewerId}/invitationStatus:
    get:
      tags:
        - "films/public/{filmId}/reviews/{reviewerId}/invitationStatus"
      summary:
        Retrieve the invitationStatus for a review that has been issued/completed
        for a film
      description:
        The invitationStatus for the review of the film with ID filmID
        issued to the user with ID reviewerId is retrieved. This operation can only
        be executed by the authenticated film owner or reviewer.
      operationId: getSingleReviewInvitationStatus
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: reviewerId
          in: path
          description: ID of the user to whom the review has been issued
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review_invitationStatus"
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviewsreviewerIdinvitationStatus
  /api/users:
    get:
      tags:
        - users
      summary: Get information about the users
      description:
        The available information (passwords excluded) about all the users
        is retrieved. This operation is available only to authenticated users.
      operationId: getUsers
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
                x-content-type: application/json
        "401":
          description: Unauthorized
          content: {}
        "404":
          description: Not Found
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: Users
...
components:
  schemas:
...
    Film:
      required:
        - owner
        - private
        - title
      type: object
      properties:
        $schema:
          type: string
          description:
            path to the JSON Schema to use for the validation of the JSON
            data structure describing the film
        id:
          type: integer
          description: unique identifier for the film data structure
        title:
          type: string
          description: textual title of the film
        owner:
          type: integer
          description: unique identifier of the the film data structure owner
        private:
          type: boolean
          description:
            "true if only the owner of the film data structure can access\
            \ it, false otherwise"
          default: true
        watchDate:
          type: string
          description: the date when the film was watched by the owner
          format: date
        rating:
          maximum: 10
          minimum: 1
          type: integer
          description: the rating the owner has given to the film
        favorite:
          type: boolean
          description:
            "true if the film is among the favourite ones of the owner,\
            \ false otherwise "
          default: false
        self:
          type: string
          format: uri
      additionalProperties: false
      description: schema for Film data structures
      example:
        owner: 6
        private: true
        watchDate: 2000-01-23
        $schema: $schema
        rating: 2
        self: http://example.com/aeiou
        id: 0
        title: title
        favorite: false
...
    inline_response_200_3:
      oneOf:
        - $ref: "#/components/schemas/Review_full"
        - $ref: "#/components/schemas/Review_public"
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT to identify the authenticated user.
      name: connect.sid
      in: cookie